Deleted AppNavigator/config folder and moved the config file into AppNavigator as config.ts
